Released Version 1/8/2014
• Clicking the “More Info” links in the Identify results dialog or
attribute table opens in an entirely new window, rather than in a dialog
superimposed on the map, allowing for printing/PDFing of the report.

Added new Freight Map viewer.

• Ability to view the user comments (dialogue section at the bottom of
the “more info” report). Had been missing due to data transfer and path
problems in old version.

• Addition of “Report” button in the attribute table, which opens a new
window with a repeated list of “more info” reports for each issue listed
in the table.

• Improved formatting of the “more info” report.

• I had to roll back James’ changes to add login and user management. It
relied on a newer version of .Net (4) than is currently configured on
our server (3.5). I’m pretty sure I can have our server updated to 4 (or
the latest, 4.5), but that will require taking it off line and
coordinating with staff in our hosting facility.